allow venus-platform to request a tunnel
As mentioned in [1] the ssh tunnel used to be started when remote support is
enabled, or remote console on VRM is enabled, which used to cover all cases.
The tunnel is used for other purposes as well nowadays and if gui-v2 is running
the tunnel is no longer needed for VRM. Since the intention is to keep
venus-access as simple as possible, this allows the tunnel to be started on
behalf of venus-platform, so only local sshd and remote support is now handled
by venus-access and any other condition can now be handled by venus-platform.

[1] https://github.com/victronenergy/venus/issues/1255